+++ hutool-poi/src/main/java/cn/hutool/poi/excel/ExcelReader.java
@@ -330,6 +330,20 @@ public class ExcelReader implements Closeable{
 	public <T> List<T> readAll(Class<T> beanType) {
 		return read(0, 1, Integer.MAX_VALUE, beanType);
 	}
+	
+	/**
+	 * 读取Excel为Bean的列表
+	 * 
+	 * @param <T> Bean类型
+	 * @param headerRowIndex 标题所在行，如果标题行在读取的内容行中间，这行做为数据将忽略，，从0开始计数
+	 * @param startRowIndex 起始行（包含，从0开始计数）
+	 * @param beanType 每行对应Bean的类型
+	 * @return Map的列表
+	 * @since 4.0.1
+	 */
+	public <T> List<T> read(int headerRowIndex, int startRowIndex, Class<T> beanType) {
+		return read(headerRowIndex, startRowIndex, Integer.MAX_VALUE, beanType);
+	}
 
 	/**
 	 * 读取Excel为Bean的列表